ADD: Styling with scss

+++ src/components/Xweet.js
@@ -1,7 +1,9 @@
 import { dbService, storageService } from 'fbase';
 import React, { useState }from 'react'
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import { faTrash, faPencilAlt } from "@fortawesome/free-solid-svg-icons";
 
-export default function Xweet({ xweetObj, isOwner }) {
+export default function Xweet({ userObj, xweetObj, isOwner }) {
     const [editing, setEditing] = useState(false)
     const [newXweet, setNewXweet] = useState(xweetObj.text)
 
@@ -9,7 +11,9 @@ export default function Xweet({ xweetObj, isOwner }) {
         const ok = window.confirm("Are you sure to delete this xweet?")
         if (ok) {
             await dbService.doc(`xweets/${xweetObj.id}`).delete();
-            await storageService.refFromURL(xweetObj.url).delete()
+            if(xweetObj.url){
+              await storageService.refFromURL(xweetObj.url).delete()
+            }
         }
     }
     const toggleEditing = () => setEditing((prev) => !prev);
@@ -27,10 +31,10 @@ export default function Xweet({ xweetObj, isOwner }) {
         setEditing(false)
     }
     return (
-      <div>
+      <div className='xweet'>
         { editing ? (
             <>
-          <form onSubmit={onSubmit}>
+          <form onSubmit={onSubmit} >
             <input 
                 type='text'
                 placeholder='New content is...'
@@ -44,15 +48,22 @@ export default function Xweet({ xweetObj, isOwner }) {
           </>
         ) : (
           <>
+            {isOwner && (
+              <div className='xweet-label'>
+                <h5>{userObj.displayName}</h5>
+                <div className='xweet-icons'>
+                  <span onClick={toggleEditing}>
+                    <FontAwesomeIcon icon={faPencilAlt} />
+                  </span>
+                  <span onClick={onDeleteClick}>
+                    <FontAwesomeIcon icon={faTrash} />
+                  </span>
+                </div>
+              </div>
+            )}
             <h4>{xweetObj.text}</h4>
             {xweetObj.url && (
-              <img src={xweetObj.url} alt='' width='50px' height='50px'/>
-            )}
-            {isOwner && (
-              <>
-                <button onClick={onDeleteClick}>Delete Xweet</button>
-                <button onClick={toggleEditing}>Edit Xweet</button>
-              </>
+              <img src={xweetObj.url} alt=''/>
             )}
           </>
         )}
